pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmalpha"
version = "0.1.0"
description = "LLM-powered cryptocurrency alpha research and backtesting framework"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Noah", email = "noah@example.com" }
]
keywords = ["crypto", "trading", "backtest", "quantitative", "research", "llm", "alpha"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "vectorbt>=0.26.0",
    "optuna>=3.0.0",
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "pyarrow>=12.0.0",
    "sqlalchemy>=2.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    # LLM providers
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
llmalpha = "llmalpha.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["llmalpha"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
